align changes made in the internal docs with public docs
