+++ docs/.vitepress/bootExamples.js
@@ -1,16 +1,15 @@
 import { buildWc } from './ex-base.js';
 
 import Alert from '../components/alert/Example.vue';
-import Attention from '../components/attention/Example.vue';
-import AttentionFlip from '../components/attention/FlipExample.vue';
-import AttentionStatic from '../components/attention/StaticExample.vue';
-import AttentionFallbackPlacements from '../components/attention/FallbackPlacementsExample.vue';
 import Badge from '../components/badge/Example.vue';
 import Box from '../components/box/Example.vue';
 import Breadcrumbs from '../components/breadcrumbs/Example.vue';
 import Button from '../components/buttons/Example.vue';
 import ButtonGroup from '../components/buttongroup/Example.vue';
 import Callout from '../components/callout/Example.vue';
+import CalloutFlip from '../components/callout/FlipExample.vue';
+import CalloutStatic from '../components/callout/StaticExample.vue';
+import CalloutFallbackPlacements from '../components/callout/FallbackPlacementsExample.vue';
 import Card from '../components/card/Example.vue';
 import Checkbox from '../components/checkbox/Example.vue';
 import Expandable from '../components/expandable/Example.vue';
@@ -18,6 +17,10 @@ import Icon from '../components/icons/Example.vue';
 import Modal from '../components/modal/Example.vue';
 import Pagination from '../components/pagination/Example.vue';
 import Pill from '../components/pill/Example.vue';
+import Popover from '../components/popover/Example.vue';
+import PopoverFlip from '../components/popover/FlipExample.vue';
+import PopoverStatic from '../components/popover/StaticExample.vue';
+import PopoverFallbackPlacements from '../components/popover/FallbackPlacementsExample.vue';
 import Radio from '../components/radio/Example.vue';
 import RadioButtons from '../components/radiobuttons/Example.vue';
 import Select from '../components/select/Example.vue';
@@ -30,22 +33,24 @@ import TextArea from '../components/textarea/Example.vue';
 import TextField from '../components/textfield/Example.vue';
 import Toast from '../components/toast/Example.vue';
 import Tooltip from '../components/tooltip/Example.vue';
+import TooltipFlip from '../components/tooltip/FlipExample.vue';
+import TooltipStatic from '../components/tooltip/StaticExample.vue';
+import TooltipFallbackPlacements from '../components/tooltip/FallbackPlacementsExample.vue';
 import Utilities from '../components/utilities/Example.vue';
 import Text from '../components/text/Example.vue';
 
 import QrColorTable from "./qr-color-table.vue";
 
 buildWc('alert-example', Alert);
-buildWc('attention-example', Attention);
-buildWc('attention-flip-example', AttentionFlip);
-buildWc('attention-static-example', AttentionStatic);
-buildWc('attention-fallback-placements-example', AttentionFallbackPlacements);
 buildWc('badge-example', Badge);
 buildWc('box-example', Box);
 buildWc('breadcrumbs-example', Breadcrumbs);
 buildWc('button-example', Button);
 buildWc('buttongroup-example', ButtonGroup);
 buildWc('callout-example', Callout);
+buildWc('callout-fallback-placements-example', CalloutFallbackPlacements);
+buildWc('callout-flip-example', CalloutFlip);
+buildWc('callout-static-example', CalloutStatic);
 buildWc('card-example', Card);
 buildWc('checkbox-example', Checkbox);
 buildWc('expandable-example', Expandable);
@@ -53,6 +58,10 @@ buildWc('icon-example', Icon);
 buildWc('modal-example', Modal);
 buildWc('pagination-example', Pagination);
 buildWc('pill-example', Pill);
+buildWc('popover-example', Popover);
+buildWc('popover-fallback-placements-example', PopoverFallbackPlacements);
+buildWc('popover-flip-example', PopoverFlip);
+buildWc('popover-static-example', PopoverStatic);
 buildWc('radio-example', Radio);
 buildWc('radiobuttons-example', RadioButtons);
 buildWc('select-example', Select);
@@ -65,6 +74,9 @@ buildWc('textarea-example', TextArea);
 buildWc('textfield-example', TextField);
 buildWc('toast-example', Toast);
 buildWc('tooltip-example', Tooltip);
+buildWc('tooltip-fallback-placements-example', TooltipFallbackPlacements);
+buildWc('tooltip-flip-example', TooltipFlip);
+buildWc('tooltip-static-example', TooltipStatic);
 buildWc('utilities-example', Utilities);
 buildWc('text-example', Text);
