--- docs/components/attention/data.json
@@ -1,25 +1,0 @@
-{
-  "title": "Attention",
-  "description": "Attention shows information that highlight content.",
-  "category": "Overlays",
-  "placeholder": {
-    "label": "Attention component"
-  },
-  "frameworks": [
-    {
-      "name": "React",
-      "status": "released"
-    },
-    {
-      "name": "Vue",
-      "status": "released"
-    },
-    {
-      "name": "Elements",
-      "status": "released"
-    },
-    {
-      "name": "Figma"
-    }
-  ]
-}
