+++ docs/components/callout/FallbackPlacementsExample.vue
@@ -1,0 +1,35 @@
+<script setup>
+import { ref } from 'vue'
+import { wAttention, wButton } from '@warp-ds/vue'
+
+const highlightTarget = ref(null)
+const highlightShowing = ref(false)
+</script>
+
+<template>
+  <div class="component">
+    <w-button
+      utility
+      :aria-expanded="highlightShowing"
+      aria-controls="highlight-example"
+      aria-details="highlight-bubbletext"
+      ref="highlightTarget"
+      @click="() => (highlightShowing = !highlightShowing)"
+    >
+      Open this dynamic highlight!
+    </w-button>
+    <w-attention
+      highlight
+      placement="bottom"
+      flip
+      cross-axis
+      :fallback-placements="['right', 'top']"
+      :target-el="highlightTarget ? highlightTarget.$el : null"
+      v-model="highlightShowing"
+      id="highlight-example"
+      tabindex="0"
+    >
+      <p id="highlight-bubbletext">This is a highlight</p>
+    </w-attention>
+  </div>
+</template>
