+++ docs/components/callout/elements.md
@@ -6,45 +6,14 @@ import '@warp-ds/elements/components/attention';
 
 ### Visual options
 
+#### Callout
+
 ```js
 <w-attention placement="right" show callout class="flex items-center">
   <div id="target" slot="target">
     <p>This is a target to callout attention element</p>
   </div>
   <span slot="message">Callout on right</span>
-</w-attention>
-```
-
-#### Tooltip
-
-```js
-<w-attention placement="right" tooltip flip>
-  <button id="target" slot="target">
-    Hover or focus to show a tooltip on right
-  </button>
-  <span slot="message">I'm a tooltip on right</span>
-</w-attention>
-```
-
-#### Popover
-
-```js
-<w-attention placement="right" popover flip cross-axis fallback-placements='["left", "bottom", "top"]'>
-  <button id="target" onclick="toggleShow()" slot="target">
-    Click to toggle a popover on bottom
-  </button>
-  <span slot="message">I'm a popover on bottom</span>
-</w-attention>
-```
-
-#### Popover with icon as target element
-
-```js
-<w-attention placement="right-end" popover distance="-8" skidding='27'>
-  <button id="target" quiet variant="utility" onclick="toggleShow()" slot="target">
-    <w-icon-info-16></w-icon-info-16>
-  </button>
-  <span slot="message">I'm a popover on right-end</span>
 </w-attention>
 ```
 
@@ -68,13 +37,13 @@ By default, the `flip` prop is set to `false`, which means that the attention co
 
 Try to scroll and see how the attention component doesn't move:
 
-<attention-static-example />
+<callout-static-example />
 
 When `flip` is set to `true`, it will instead trigger [Floating-ui's flip() function](https://floating-ui.com/docs/flip)  that will make sure that the attention component stays in viewport, by flipping it to the opposite side.
 
 Try to scroll and see how the attention component moves its position to the opposite side to keep itself in viewport as long as possible:
 
-<attention-flip-example />
+<callout-flip-example />
 
 ### Cross-axis prop
 The `cross-axis` prop decides whether to check for cross axis overflow or not when `flip` is set to `true`.
@@ -89,7 +58,7 @@ Read more: [Floating-ui fallbackPlacements](https://floating-ui.com/docs/flip#fa
 
 Try to scroll and see how the attention component's position starts at the `bottom` but then moves to the `right` and then to the `top`: 
 
-<attention-fallback-placements-example />
+<callout-fallback-placements-example />
 
 ### Accessibility
 If the Attention element has "left" or "top" position, it should be placed before the target element in the DOM.
@@ -102,13 +71,13 @@ To do that set the `role` attribute on the relevant text element nested in `w-at
 The `aria-details` attribute is on the target element, not on `w-attention`.
 
 ```js
-<w-attention placement="top" tooltip>
+<w-attention placement="top" highlight>
   <div slot="message">
-    <p id="aria-content" role="tooltip">This tooltip text is important</p>
+    <p id="aria-content">This highlight text is important</p>
     <p>(this text is less relevant)</p>
   </div>
   <button aria-describedby="aria-content" id="target" slot="target">
-    Click to toggle a tooltip on top
+    Click to toggle a highlight on top
   </button>
 </w-attention>
 ```
