+++ docs/components/callout/vue.md
@@ -40,117 +40,6 @@ const showing = ref(false);
 </template>
 ```
 
-#### Tooltip
-
-```vue
-<script setup>
-import { ref } from 'vue';
-import { wAttention, wButton } from '#components';
-
-const target = ref(null);
-const showing = ref(false);
-</script>
-<template>
-  <w-button
-    utility
-    ref="target"
-    aria-describedby="tooltip-bubbletext"
-    :aria-expanded="showing"
-    @mouseenter="
-      showing = true
-      target = $refs.target
-    "
-    @mouseleave="showing = false"
-    @keydown.escape="showing = false"
-    @focus="showing = true"
-    @blur="showing = false"
-  >
-    Hover over me
-  </w-button>
-  <w-attention
-    tooltip
-    placement="right"
-    flip
-    :target-el="target?.$el"
-    v-model="showing"
-  >
-    <p id="tooltip-bubbletext">Hello Warp!</p>
-  </w-attention>
-</template>
-```
-
-#### Popover
-
-```vue
-<script setup>
-import { ref } from 'vue';
-import { wAttention, wButton } from '#components';
-
-const target = ref(null);
-const showing = ref(false);
-</script>
-<template>
-  <w-button
-    utility
-    :aria-expanded="showing"
-    aria-controls="popover-example"
-    aria-details="popover-bubbletext"
-    ref="target"
-    @click="() => (showing = !showing)"
-  >
-    Open popover
-  </w-button>
-  <w-attention
-    popover
-    placement="right"
-    flip
-    cross-axis
-    :fallback-placements="['left', 'bottom', 'top']"
-    :target-el="target ? target.$el : null"
-    v-model="showing"
-    id="popover-example"
-  >
-    <p id="popover-bubbletext">Hello Warp!</p>
-  </w-attention>
-</template>
-```
-
-#### Popover with icon as target element
-
-```vue
-<script setup>
-import { ref } from 'vue';
-import { wAttention, wButton } from '#components';
-import IconInfo16 from '@warp-ds/icons/vue/info-16';
-
-const target = ref(null);
-const showing = ref(false);
-</script>
-<template>
- <w-button
-  :aria-expanded="popoverIconTargetShowing"
-  aria-controls="popover-icon-target-example"
-  type="button"
-  utility
-  quiet
-  ref="popoverIconTarget"
-  @click="() => (popoverIconTargetShowing = !popoverIconTargetShowing)"
-  >
-    <icon-info-16 />
-  </w-button>
-  <w-attention
-    popover
-    placement="right-end"
-    :distance="-8"
-    :skidding="27"
-    :target-el="popoverIconTarget ? popoverIconTarget.$el : null"
-    v-model="popoverIconTargetShowing"
-  >
-    <p>Hello Warp!</p>
-  </w-attention>
-</template>
-```
-
 #### Highlight (with optional close button)
 
 ```vue
@@ -191,13 +80,13 @@ By default, the `flip` prop is set to `false`, which means that the attention co
 
 Try to scroll and see how the attention component doesn't move:
 
-<attention-static-example />
+<callout-static-example />
 
 When `flip` is set to `true`, it will instead trigger [Floating-ui's flip() function](https://floating-ui.com/docs/flip) that will make sure that the attention component stays in viewport, by flipping it to the opposite side.
 
 Try to scroll and see how the attention component moves its position to the opposite side to keep itself in viewport as long as possible:
 
-<attention-flip-example />
+<callout-flip-example />
 
 ### Cross-axis prop
 The `cross-axis` prop decides whether to check for cross axis overflow or not when `flip` is set to `true`.
@@ -213,7 +102,7 @@ Read more: [Floating-ui fallbackPlacements](https://floating-ui.com/docs/flip#fa
 
 Try to scroll and see how the attention component's position starts at the `bottom` but then moves to the `right` and then to the `top`: 
 
-<attention-fallback-placements-example />
+<callout-fallback-placements-example />
 
 ### Accessibility
 The attention component handles accessibility automatically by wrapping its slotted content with a `div` that has a default `role` attribute (`role="tooltip"` for tooltip and `role="img"` otherwise), and a default localized `aria-label`.
@@ -222,14 +111,14 @@ It is possible to remove the default `role` and override the `aria-label` attrib
 
 ```vue
 <w-attention
-  tooltip
+  highlight
   placement='bottom'
-  v-model='tooltipShowing'
+  v-model='highlightShowing'
   role=''
   aria-label=''
 >
-  <p id='tooltip-bubbletext' role='tooltip'>
-    I'm a tooltip speech bubble with overridden role and aria-label attributes pointing up.
+  <p id='highlight-bubbletext'>
+    I'm a highlight speech bubble with overridden role and aria-label attributes pointing up.
   </p>
 </w-attention>
 ```
