+++ docs/components/popover/FlipExample.vue
@@ -25,6 +25,7 @@ const popoverShowing = ref(false)
       :target-el="popoverTarget ? popoverTarget.$el : null"
       v-model="popoverShowing"
       id="popover-example"
+      tabindex="0"
     >
       <p id="popover-bubbletext">This is a popover</p>
     </w-attention>
