+++ docs/components/popover/StaticExample.vue
@@ -24,6 +24,7 @@ const popoverShowing = ref(false)
       :target-el="popoverTarget ? popoverTarget.$el : null"
       v-model="popoverShowing"
       id="popover-example"
+      tabindex="0"
     >
       <p id="popover-bubbletext">This is a popover</p>
     </w-attention>
