+++ docs/components/popover/index.md
@@ -15,7 +15,7 @@
 
 ## Example
 <ThemeSwitcher />
-<callout-example />
+<popover-example />
 
 ## Usage
