+++ docs/components/popover/vue.md
@@ -21,64 +21,6 @@ import { wAttention } from '@warp-ds/vue/attention';
 
 ### Visual options
 
-#### Callout
-
-```vue
-<script setup>
-import { ref } from 'vue';
-import { wAttention, wBox } from '#components';
-
-const showing = ref(false);
-</script>
-<template>
-  <w-box neutral as="h4" aria-details="callout-bubbletext">
-    I am a box full of info
-  </w-box>
-  <w-attention callout placement="right" v-model="showing">
-    <p id="callout-bubbletext">Hello Warp! This thing is new!</p>
-  </w-attention>
-</template>
-```
-
-#### Tooltip
-
-```vue
-<script setup>
-import { ref } from 'vue';
-import { wAttention, wButton } from '#components';
-
-const target = ref(null);
-const showing = ref(false);
-</script>
-<template>
-  <w-button
-    utility
-    ref="target"
-    aria-describedby="tooltip-bubbletext"
-    :aria-expanded="showing"
-    @mouseenter="
-      showing = true
-      target = $refs.target
-    "
-    @mouseleave="showing = false"
-    @keydown.escape="showing = false"
-    @focus="showing = true"
-    @blur="showing = false"
-  >
-    Hover over me
-  </w-button>
-  <w-attention
-    tooltip
-    placement="right"
-    flip
-    :target-el="target?.$el"
-    v-model="showing"
-  >
-    <p id="tooltip-bubbletext">Hello Warp!</p>
-  </w-attention>
-</template>
-```
-
 #### Popover
 
 ```vue
@@ -151,53 +93,19 @@ const showing = ref(false);
 </template>
 ```
 
-#### Highlight (with optional close button)
-
-```vue
-<script setup>
-import { ref } from 'vue'
-import { wAttention, wButton } from '#components'
-
-const highlightTarget = ref(null)
-const highlightShowing = ref(false)
-</script>
-<template>
-  <w-button
-    :aria-expanded="highlightShowing"
-    aria-controls="highlight-attention-example"
-    aria-details="highlighted-bubbletext"
-    ref="highlightTarget"
-    @click="highlightShowing = !highlightShowing"
-    >Click me</w-button
-  >
-  <w-attention
-    id="highlight-attention-example"
-    highlight
-    placement="bottom"
-    flip
-    :fallback-placements="['top']"
-    can-close
-    @dismiss="highlightShowing = false"
-    :target-el="highlightTarget ? highlightTarget.$el : null"
-    v-model="highlightShowing"
-  >
-    <p id="highlighted-bubbletext">I'm a dismissible highlight</p>
-  </w-attention>
-</template>
-```
 ### Flip prop
 The attention component uses the Floating-ui library to calculate its position.
 By default, the `flip` prop is set to `false`, which means that the attention component will not flip its position to the opposite side.
 
 Try to scroll and see how the attention component doesn't move:
 
-<attention-static-example />
+<popover-static-example />
 
 When `flip` is set to `true`, it will instead trigger [Floating-ui's flip() function](https://floating-ui.com/docs/flip) that will make sure that the attention component stays in viewport, by flipping it to the opposite side.
 
 Try to scroll and see how the attention component moves its position to the opposite side to keep itself in viewport as long as possible:
 
-<attention-flip-example />
+<popover-flip-example />
 
 ### Cross-axis prop
 The `cross-axis` prop decides whether to check for cross axis overflow or not when `flip` is set to `true`.
@@ -213,7 +121,7 @@ Read more: [Floating-ui fallbackPlacements](https://floating-ui.com/docs/flip#fa
 
 Try to scroll and see how the attention component's position starts at the `bottom` but then moves to the `right` and then to the `top`: 
 
-<attention-fallback-placements-example />
+<popover-fallback-placements-example />
 
 ### Accessibility
 The attention component handles accessibility automatically by wrapping its slotted content with a `div` that has a default `role` attribute (`role="tooltip"` for tooltip and `role="img"` otherwise), and a default localized `aria-label`.
@@ -222,14 +130,14 @@ It is possible to remove the default `role` and override the `aria-label` attrib
 
 ```vue
 <w-attention
-  tooltip
+  popover
   placement='bottom'
-  v-model='tooltipShowing'
+  v-model='popoverShowing'
   role=''
   aria-label=''
 >
-  <p id='tooltip-bubbletext' role='tooltip'>
-    I'm a tooltip speech bubble with overridden role and aria-label attributes pointing up.
+  <p id='popover-bubbletext'>
+    I'm a popover speech bubble with overridden role and aria-label attributes pointing up.
   </p>
 </w-attention>
 ```
