+++ docs/components/tooltip/elements.md
@@ -26,55 +26,19 @@ import '@warp-ds/elements/components/attention';
 </w-attention>
 ```
 
-#### Popover
-
-```js
-<w-attention placement="right" popover flip cross-axis fallback-placements='["left", "bottom", "top"]'>
-  <button id="target" onclick="toggleShow()" slot="target">
-    Click to toggle a popover on bottom
-  </button>
-  <span slot="message">I'm a popover on bottom</span>
-</w-attention>
-```
-
-#### Popover with icon as target element
-
-```js
-<w-attention placement="right-end" popover distance="-8" skidding='27'>
-  <button id="target" quiet variant="utility" onclick="toggleShow()" slot="target">
-    <w-icon-info-16></w-icon-info-16>
-  </button>
-  <span slot="message">I'm a popover on right-end</span>
-</w-attention>
-```
-
-#### Highlight (with optional close button)
-
-```js
-<w-attention placement="right" highlight id="highlight" can-close flip fallback-placements='["top"]'>
-  <button
-    id="highlightTarget"
-    slot="target"
-  >
-    Click to toggle a highlight on right
-  </button>
-  <span slot="message">I'm a dismissible highlight on right</span>
-</w-attention>
-```
-
 ### Flip prop
 The attention component uses the Floating-ui library to calculate its position.
 By default, the `flip` prop is set to `false`, which means that the attention component will not flip its position to the opposite side.
 
 Try to scroll and see how the attention component doesn't move:
 
-<attention-static-example />
+<tooltip-static-example />
 
 When `flip` is set to `true`, it will instead trigger [Floating-ui's flip() function](https://floating-ui.com/docs/flip)  that will make sure that the attention component stays in viewport, by flipping it to the opposite side.
 
 Try to scroll and see how the attention component moves its position to the opposite side to keep itself in viewport as long as possible:
 
-<attention-flip-example />
+<tooltip-flip-example />
 
 ### Cross-axis prop
 The `cross-axis` prop decides whether to check for cross axis overflow or not when `flip` is set to `true`.
@@ -87,9 +51,9 @@ By default, `fallback-placements` is `undefined`. If `flip`is set to `true`, the
 
 Read more: [Floating-ui fallbackPlacements](https://floating-ui.com/docs/flip#fallbackplacements)
 
-Try to scroll and see how the attention component's position starts at the `bottom` but then moves to the `right` and then to the `top`: 
+Try to scroll and see how the attention component's position starts at the `bottom` but then moves to the `right` but then moves to the `top`: 
 
-<attention-fallback-placements-example />
+<tooltip-fallback-placements-example />
 
 ### Accessibility
 If the Attention element has "left" or "top" position, it should be placed before the target element in the DOM.
