+++ docs/components/tooltip/react.md
@@ -12,21 +12,6 @@ import { Attention } from '@warp-ds/react/components/attention'
 
 ### Visual options
 
-#### Callout
-
-```js
-<div>
-  <Box info>
-    <h1 aria-details='callout-bubbletext'>I am a box full of info</h1>
-  </Box>
-  <Attention callout placement='right' isShowing={true}>
-    <p id='callout-bubbletext'>
-      I'm a callout because that box over there is new or something
-    </p>
-  </Attention>
-</div>
-```
-
 #### Tooltip
 
 ```js
@@ -62,169 +47,19 @@ function Example() {
 }
 ```
 
-#### Popover
-
-```js
-function Example() {
-  const [show, setShow] = React.useState(false)
-  const targetEl = React.useRef()
-
-  return (
-    <div>
-      <Button
-        small
-        aria-expanded={show}
-        aria-controls='popover-attention-example'
-        aria-details='popover-bubbletext'
-        utility
-        onClick={() => setShow(!show)}
-        ref={targetEl}
-      >
-        Open menu
-      </Button>
-      <Attention
-        popover
-        placement='right'
-        flip={true}
-        crossAxis={true}
-        fallbackPlacements={['left', 'bottom', 'top']}
-        targetEl={targetEl}
-        isShowing={show}
-        id='popover-attention-example'
-      >
-        <ul id='popover-bubbletext'>
-          <li>Hello</li>
-          <li>World</li>
-        </ul>
-      </Attention>
-    </div>
-  )
-}
-```
-
-#### Popover with icon as target element
-
-```js
-function PopoverIconAsTargetEl() {
-  const [show, setShow] = React.useState(false)
-  const containerRef = React.useRef<HTMLDivElement>(null)
-  const targetEl = React.useRef<any>(null);
-
-  React.useEffect(() => {
-    function onBlurHandler(e) {
-      if (containerRef.current && !containerRef.current.contains(e.target)) {
-        setShow(false)
-      }
-    }
-    document.addEventListener('mousedown', onBlurHandler)
-    return () => {
-      document.removeEventListener('mousedown', onBlurHandler)
-    }
-  })
-
-  return (
-  <div className='flex items-center justify-center'>
-    <div ref={containerRef}>
-      <Button
-        small
-        aria-expanded={show}
-        aria-controls='popover-icon-target-el-attention-example'
-        utility
-        quiet
-        onClick={() => setShow(!show)}
-        className='w-max mb-0'
-        ref={targetEl}
-      >
-        <IconInfo16 />
-      </Button>
-      <Attention
-        popover
-        placement='right-end'
-        distance={-8}
-        skidding={27}
-        targetEl={targetEl}
-        isShowing={show}
-      >
-        <ul className='bg-white w-full text-center'>
-          <li
-            tabIndex={0}
-            className='p-2 px-40 hover:bg-blue-200 hover:text-blue-500 cursor-pointer'
-          >
-            Hello
-          </li>
-          <li
-            tabIndex={0}
-            className='p-2 px-40 hover:bg-blue-200 hover:text-blue-500 cursor-pointer'
-          >
-            World
-          </li>
-        </ul>
-      </Attention>
-    </div>
-  </div>
-  )
-}
-```
-
-#### Highlight (with optional close button)
-
-```js
-function DismissibleHighlight() {
-  const [show, setShow] = React.useState(false)
-  const targetEl = React.useRef()
-
-  return (
-    <div className='flex flex-col justify-between h-[200]'>
-      <Button
-        small
-        aria-expanded={show}
-        aria-controls='highlight-attention-example'
-        utility
-        onClick={() => setShow(!show)}
-        className='w-max mb-0'
-      >
-        Show an onboarding hint
-      </Button>
-      <div>
-        <div ref={targetEl} className="w-2/3" aria-details='highlight-bubble-text'>
-          <Box info>
-            <h1>I am a box full of info</h1>
-          </Box>
-        </div>
-        <Attention
-          highlight
-          canClose
-          onDismiss={() => setShow(false)}
-          placement='bottom'
-          isShowing={show}
-          targetEl={targetEl}
-          flip={true}
-          fallbackPlacements={['top']}
-          id='highlight-attention-example'
-        >
-          <p id='highlight-bubble-text'>
-            I'm a highlight that can dismiss itself
-          </p>
-        </Attention>
-      </div>
-    </div>
-  )
-}
-```
-
 ### Flip prop
 The attention component uses the Floating-ui library to calculate its position.
 By default, the `flip` prop is set to `false`, which means that the attention component will not flip its position to the opposite side.
 
 Try to scroll and see how the attention component doesn't move:
 
-<attention-static-example />
+<tooltip-static-example />
 
 When `flip` is set to `true`, it will instead trigger [Floating-ui's flip() function](https://floating-ui.com/docs/flip) that will make sure that the attention component stays in viewport, by flipping it to the opposite side.
 
 Try to scroll and see how the attention component moves its position to the opposite side to keep itself in viewport as long as possible:
 
-<attention-flip-example />
+<tooltip-flip-example />
 
 ### CrossAxis prop
 The `crossAxis` prop decides whether to check for cross axis overflow or not when `flip` is set to `true`.
@@ -238,9 +73,9 @@ If `flip`is set to `true`, then you have the option to also use `fallbackPlaceme
 
 Read more: [Floating-ui fallbackPlacements](https://floating-ui.com/docs/flip#fallbackplacements)
 
-Try to scroll and see how the attention component's position starts at the `bottom` but then moves to the `right` and then to the `top`: 
+Try to scroll and see how the attention component's position starts at the `right` but then moves to the `top`: 
 
-<attention-fallback-placements-example />
+<tooltip-fallback-placements-example />
 
 ### Accessibility
 The attention component handles accessibility automatically by wrapping its slotted content with a `div` that has a default `role` attribute (`role="tooltip"` for tooltip and `role="img"` otherwise), and a default localized `aria-label`.
